server/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
django-environ==0.12.1
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
Pillow==11.1.0
